+++ src/main/java/com/lowdragmc/photon/client/gameobject/emitter/data/number/color/RandomGradient.java
@@ -43,10 +43,10 @@ public class RandomGradient implements NumberFunction {
     }
 
     @Override
-    public Float get(float t, Supplier<Float> lerp) {
+    public Integer get(float t, Supplier<Float> lerp) {
         int color0 = gradientColor0.getColor(t);
         int color1 = gradientColor1.getColor(t);
-        return (float) ColorUtils.blendColor(color0, color1, lerp.get());
+        return ColorUtils.blendColor(color0, color1, lerp.get());
     }
 
     @Override
